feat(multimode): Add initial no Drawer Mode
add preferences, utils and all apps task

[SahilSonar - adapted to A14]

+++ src/com/android/launcher3/SessionCommitReceiver.java
@@ -35,6 +35,8 @@ import com.android.launcher3.util.Executors;
 
 import java.util.Locale;
 
+import foundation.e.bliss.multimode.MultiModeController;
+
 /**
  * BroadcastReceiver to handle session commit intent.
  */
@@ -69,16 +71,16 @@ public class SessionCommitReceiver extends BroadcastReceiver {
         boolean alreadyAddedPromiseIcon =
                 packageInstallerCompat.promiseIconAddedForId(info.getSessionId());
         if (TextUtils.isEmpty(info.getAppPackageName())
-                || info.getInstallReason() != PackageManager.INSTALL_REASON_USER
+                || (info.getInstallReason() != PackageManager.INSTALL_REASON_USER && !isSingleLayer())
                 || alreadyAddedPromiseIcon) {
             FileLog.d(LOG,
                     String.format(Locale.ENGLISH,
                             "Removing PromiseIcon for package: %s, install reason: %d,"
-                            + " alreadyAddedPromiseIcon: %s",
-                    info.getAppPackageName(),
-                    info.getInstallReason(),
-                    alreadyAddedPromiseIcon
-                )
+                                    + " alreadyAddedPromiseIcon: %s",
+                            info.getAppPackageName(),
+                            info.getInstallReason(),
+                            alreadyAddedPromiseIcon
+                    )
             );
             packageInstallerCompat.removePromiseIconId(info.getSessionId());
             return;
@@ -104,6 +106,11 @@ public class SessionCommitReceiver extends BroadcastReceiver {
                 && UserCache.getInstance(context).getUserInfo(user).isPrivate()) {
             return false;
         }
-        return LauncherPrefs.getPrefs(context).getBoolean(ADD_ICON_PREFERENCE_KEY, true);
+        return LauncherPrefs.getPrefs(context).getBoolean(ADD_ICON_PREFERENCE_KEY, true)
+                || isSingleLayer();
+    }
+
+    public static boolean isSingleLayer() {
+        return MultiModeController.isSingleLayerMode();
     }
 }
